now handles all data fetching for users, groups and creation form update: renders two childs CreationForm and RowsHandler
